SOLR-17806 Move SolrMetricTestUtils to test-framework
